requirements.txt
# core packages
numpy==1.24.4
transformers==4.37.2
torch>=2
torchmetrics==1.3.1
tokenizers==0.15.2
datasets==2.17.0
ai2-tango[torch,wandb,transformers]==1.3.2
allennlp-light==1.0.0

# other
conllu==4.5.3
black==24.2.0
isort==5.13.2
psutil==5.9.8
scikit-learn==1.4.1.post1
PyGithub==2.2.0
tensorboard==2.16.1
stanza==1.7.0
